+++ Gemfile
@@ -1,4 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'cli-ui', '~>1.3.0'
-gem 'rubocop', '~>0.81.0'
+gem 'cli-ui', '~>1.3'
+gem 'curses', '~> 1.3'
+gem 'rubocop', '~>0.81'
+gem 'strings', '~> 0.1'
